fix(ui): resolve autoview display failures and standardize preview state - Unified redundant preview mode variables to prevent synchronization drift. - Corrected file list indexing for Autoview to account for absolute scroll offsets. - Integrated sorting into list construction to ensure consistent UI sourcing. - Resolved path handling and type conversion issues in file copy operations. - Modernized legacy type definitions and refined global window refresh logic.

+++ src/fs/freesp.c
@@ -5,101 +5,105 @@
  *
  ***************************************************************************/
 
-
 #include "ytree.h"
+#include <unistd.h>
 
 /* Removed #include <dos.h> and #include <hurd/hurd_types.h> */
 
 /* Volume-Name und freien Plattenplatz ermitteln */
 /*-----------------------------------------------*/
 
-int GetDiskParameter( char *path,
-		      char *volume_name,
-		      LONGLONG *avail_bytes,
-		      LONGLONG *total_disk_space,
-		      Statistic *s
-		    )
-{
+int GetDiskParameter(char *path, char *volume_name, long long *avail_bytes,
+                     long long *total_disk_space, Statistic *s) {
 
 #ifdef WIN32
   struct _diskfree_t diskspace;
 #else
   /* Use struct statvfs as the modern POSIX standard */
   struct statvfs statfs_struct;
-  #define FS_STRUCT_IS_STATVFS
+#define FS_STRUCT_IS_STATVFS
 #endif
 
   char *p;
   char *fname;
-  int  result;
-  LONGLONG bfree;
-  LONGLONG this_disk_space;
-
+  int result;
+  long long bfree;
+  long long this_disk_space;
 
 #ifdef WIN32
-  if( ( result = _getdiskfree( 0, &diskspace ) ) == 0 )
+  if ((result = _getdiskfree(0, &diskspace)) == 0)
 #else
   /* Use the STATFS macro defined in ytree.h which resolves to statvfs */
-  if( ( result = STATFS( path, &statfs_struct, 0, 0 ) ) == 0 )
+  if ((result = STATFS(path, &statfs_struct, 0, 0)) == 0)
 #endif /* WIN32 */
   {
-    if( volume_name )
-    {
+    if (volume_name) {
       /* Name ermitteln */
       /*----------------*/
 
       /* Renamed usage: s->mode -> s->login_mode */
-      if( s->login_mode == DISK_MODE || s->login_mode == USER_MODE )
-      {
+      if (s->login_mode == DISK_MODE || s->login_mode == USER_MODE) {
 
 #if defined(__linux__)
-	/* Minimal Linux/GNU FS Type Detection */
-	switch( statfs_struct.f_type ) {
-	  case 0xEF53: fname = "EXT2/3/4"; break;
-	  case 0x9660: fname = "ISOFS"; break;
-	  case 0x4d44: fname = "DOS/FAT"; break;
-	  case 0x6969: fname = "NFS"; break;
-	  case 0x9fa0: fname = "PROC"; break;
-      case 0x5346544e: /* 'NTFS' magic from ntfs-3g */
-	  case 0x53465435: fname = "NTFS"; break;
-      case 0x72656d6f: fname = "WSLFS/9P"; break; /* For WSL2 mounts */
-	  default: fname = "OTHER-FS";
-	}
+        /* Minimal Linux/GNU FS Type Detection */
+        switch (statfs_struct.f_type) {
+        case 0xEF53:
+          fname = "EXT2/3/4";
+          break;
+        case 0x9660:
+          fname = "ISOFS";
+          break;
+        case 0x4d44:
+          fname = "DOS/FAT";
+          break;
+        case 0x6969:
+          fname = "NFS";
+          break;
+        case 0x9fa0:
+          fname = "PROC";
+          break;
+        case 0x5346544e: /* 'NTFS' magic from ntfs-3g */
+        case 0x53465435:
+          fname = "NTFS";
+          break;
+        case 0x72656d6f:
+          fname = "WSLFS/9P";
+          break; /* For WSL2 mounts */
+        default:
+          fname = "OTHER-FS";
+        }
 #elif defined(__APPLE__)
-	/* Rely on statvfs f_fstypename */
-	fname = statfs_struct.f_fstypename;
+        /* Rely on statvfs f_fstypename */
+        fname = statfs_struct.f_fstypename;
 #elif defined(__NetBSD__) || defined(__OpenBSD__) || defined(__FreeBSD__)
-	/* BSDs often provide f_fstypename via statvfs */
-	fname = statfs_struct.f_fstypename;
+        /* BSDs often provide f_fstypename via statvfs */
+        fname = statfs_struct.f_fstypename;
 #else
         /* Generic POSIX fallback (includes WSL/other Unix-likes) */
         fname = "UNIX-FS";
 #endif
 
-        (void) strncpy( volume_name,
-	                fname,
-		        MINIMUM( DISK_NAME_LENGTH, strlen( fname ) )
-		      );
-        volume_name[ MINIMUM( DISK_NAME_LENGTH, strlen( fname ))] = '\0';
-      }
-      else
-      {
+        (void)strncpy(volume_name, fname,
+                      MINIMUM(DISK_NAME_LENGTH, strlen(fname)));
+        volume_name[MINIMUM(DISK_NAME_LENGTH, strlen(fname))] = '\0';
+      } else {
         /* ARCHIVE_MODE */
         /*--------------*/
 
-        if( ( p = strrchr( s->login_path, FILE_SEPARATOR_CHAR ) ) == NULL )
+        if ((p = strrchr(s->login_path, FILE_SEPARATOR_CHAR)) == NULL)
           p = s->login_path;
-        else p++;
+        else
+          p++;
 
-        (void) strncpy( volume_name, p, sizeof( s->disk_name ) );
-        volume_name[sizeof( s->disk_name )] = '\0';
+        (void)strncpy(volume_name, p, sizeof(s->disk_name));
+        volume_name[sizeof(s->disk_name)] = '\0';
       }
     } /* volume_name */
 
 #ifdef WIN32
-    *avail_bytes = (LONGLONG) diskspace.bytes_per_sector *
-                   (LONGLONG) diskspace.sectors_per_cluster *
-                   (LONGLONG) diskspace.avail_clusters;
+    *avail_bytes = (long long)diskspace.bytes_per_sector *
+                   (long long)diskspace.sectors_per_cluster *
+                   (long long)diskspace.avail_clusters;
     this_disk_space = 900000000L; /* for now.. */
 #else
     /* Consolidated POSIX logic for disk space */
@@ -107,34 +111,24 @@ int GetDiskParameter( char *path,
     /* Use f_bavail for non-root, f_bfree for root/fallback */
     bfree = getuid() ? statfs_struct.f_bavail : statfs_struct.f_bfree;
 
-    if( bfree < 0L ) bfree = 0L;
+    if (bfree < 0L)
+      bfree = 0L;
 
-    /* statvfs uses f_frsize (fragment size) */
-    #define FRAG_SIZE statfs_struct.f_frsize
+/* statvfs uses f_frsize (fragment size) */
+#define FRAG_SIZE statfs_struct.f_frsize
 
     *avail_bytes = bfree * FRAG_SIZE;
-    this_disk_space   = (LONGLONG)statfs_struct.f_blocks * FRAG_SIZE;
+    this_disk_space = (long long)statfs_struct.f_blocks * FRAG_SIZE;
 
 #endif /* WIN32 */
 
-    if( total_disk_space )
-    {
+    if (total_disk_space) {
       *total_disk_space = this_disk_space;
     }
   }
-  return( result );
+  return (result);
 }
 
-
-
-
-int GetAvailBytes(LONGLONG *avail_bytes, Statistic *s)
-{
-  return( GetDiskParameter( s->tree->name,
-			    NULL,
-			    avail_bytes,
-			    NULL,
-			    s
-			  )
-        );
+int GetAvailBytes(long long *avail_bytes, Statistic *s) {
+  return (GetDiskParameter(s->tree->name, NULL, avail_bytes, NULL, s));
 }
